fix: add deprecated zeebe property to condition
Legacy config postprocessing was not backported, thus
the condition needs to consider both properties.
Added tests to cover usage of different enable properties.
